구조 변경을 위한 임시 저장
1. 과도한 설계로 인한 기획 변경.
1-1. 메인 스탯 변경 예정.

+++ Assets/2.Scripts/Contents/Entity/module/ResourceModule.cs
@@ -19,7 +19,7 @@ public class ResourceModule
 
     public ResourceModule(ActorEntitySo so)
     {
-        foreach (KeyValuePair<MainStatType, float> kv in so.MainStatDic)
+        foreach (KeyValuePair<MainStatType, int> kv in so.MainStatDic)
         {
             MainStat main = new MainStat(kv.Key, kv.Value);
